[BUGFIX] Add page uid to link browser tabs for correct TSconfig loading
This is a follow up to ea4d2a5. Because of a missing page uid
parameter in links to other linkHandler types the ConditionMatcher
could not evaluate conditions in INCLUDE_TYPOSCRIPT that rely on the
page uid, like PIDinRootline, if the user changes the type in the
link browser modal.